UID check from token

+++ flowkit_ui_backend/apis/accounts_api.py
@@ -71,6 +71,11 @@ async def delete_user(
         raise NotImplementedError("The /delete_user endpoint is not yet implemented")
 
     try:
+        if uid != token_auth0.sub:
+            return JSONResponse(
+                status_code=HTTPStatus.UNAUTHORIZED,
+                content="You can only request user data on yourself"
+            )
         logger.debug("Starting request")
         impl_result = await accounts_api_impl.delete_user(
             uid, pool=request.app.state.pool, token_model=token_auth0
